requirements.txt
Flask==3.0.0
openai==1.51.0
python-dotenv==1.0.1
# Pin httpx for openai client compatibility
httpx==0.27.2
gunicorn==22.0.0
